refactor(test): Migrate test framework to jest
https://zanata.atlassian.net/browse/ZNTA-1202
