+++ .travis.yml
@@ -1,3 +1,3 @@
 language: node_js
 node_js:
-  - "4.2"
+  - "5.6.0"
