+++ package.json
@@ -14,8 +14,7 @@
   "scripts": {
     "watch": "cd app && webpack-dev-server -d --progress --config webpack.dev.config.js --inline --hot --content-base build/",
     "build": "cd app && webpack --progress --config webpack.prod.config.js",
-    "test": "babel-tape-runner \"test/**/*.jsx\" | faucet",
-    "test-raw": "babel-tape-runner \"test/**/*.jsx\"",
+    "test": "jest",
     "storybook": "start-storybook -p 9001 -s ./app/build",
     "build-storybook": "build-storybook -s ./app/build",
     "spritesheet": "gulp icons",
@@ -26,8 +25,8 @@
     "url": "http://github.com/zanata/zanata-spa.git"
   },
   "engines": {
-    "node": ">=5.4.1",
-    "npm": ">=3.3.12"
+    "node": "5.6.0",
+    "npm": "3.6.0"
   },
   "license": "LGPL-2.1",
   "devDependencies": {
@@ -35,6 +34,7 @@
     "autoprefixer": "6.3.6",
     "babel-core": "6.9.0",
     "babel-eslint": "4.1.8",
+    "babel-jest": "13.0.0",
     "babel-loader": "6.2.4",
     "babel-plugin-react-intl": "2.1.3",
     "babel-preset-es2015": "6.9.0",
@@ -65,9 +65,11 @@
     "gulp-svg-sprite": "1.3.1",
     "gulp-webserver": "0.4.0",
     "isomorphic-fetch": "2.2.1",
+    "jest-cli": "13.0.0",
     "lodash": "3.10.1",
     "postcss-loader": "0.8.2",
     "react": "0.14.8",
+    "react-addons-test-utils": "0.14.8",
     "react-addons-update": "0.14.8",
     "react-dom": "0.14.8",
     "react-intl": "2.1.2",
@@ -80,7 +82,6 @@
     "redux-simple-router": "2.0.4",
     "redux-thunk": "1.0.3",
     "reselect": "1.1.0",
-    "retap": "2.0.0",
     "rework-calc": "1.1.0",
     "rework-color-function": "1.2.1",
     "rework-custom-media": "0.1.1",
@@ -105,5 +106,20 @@
     "suitcss-utils-size": "0.7.1",
     "suitcss-utils-text": "0.4.1",
     "text-diff": "1.0.1"
+  },
+  "jest": {
+    "scriptPreprocessor": "./node_modules/babel-jest",
+    "unmockedModulePathPatterns": [
+      "./node_modules/react",
+      "./node_modules/react-dom",
+      "./node_modules/react-addons-test-utils",
+      "./node_modules/fbjs",
+      "./node_modules/lodash",
+      "./app"
+    ],
+    "moduleFileExtensions": [
+      "js",
+      "jsx"
+    ]
   }
 }
