--- test/components/ProjectVersionLinkTest.jsx
@@ -1,26 +1,0 @@
-import React from 'react'
-import ProjectVersionLink from '../../app/components/ProjectVersionLink'
-import test from 'retap'
-
-test('ProjectVersionLink markup', function ProjectVersionLinkMarkup (t) {
-
-  const actual = <ProjectVersionLink project={{name: 'Weight Gain'}}
-    version="4000"
-    url="https://en.wikipedia.org/wiki/Weight_Gain_4000"/>
-
-  const expected = (
-    <a href="https://en.wikipedia.org/wiki/Weight_Gain_4000"
-       className="Link--invert Header-item u-inlineBlock">
-      <span className="u-sPH-1-4 u-sizeWidth1 u-gtemd-hidden">
-        <i className="i i--arrow-left"></i>
-      </span>
-      <span className="Editor-currentProject u-sm-hidden u-sML-1-2">
-        <span>Weight Gain</span> <span
-          className="u-textMuted">4000</span>
-      </span>
-    </a>
-  )
-
-  t.isSameMarkup(actual, expected)
-  t.end()
-})
